Fix usage of TestPlan API in instrumentation tests to match binary-incompatible change in 5.12
